removed old, changed to new hand eye
